+++ example/add/hook/build.dart
@@ -2,7 +2,7 @@
 import 'dart:io';
 
 import 'package:add/src/hook_helpers/hook_helpers.dart';
-import 'package:native_assets_cli/native_assets_cli.dart';
+import 'package:hooks/hooks.dart';
 import 'package:native_toolchain_cmake/native_toolchain_cmake.dart';
 
 void main(List<String> args) async {