+++ lib/src/builder/run_builder.dart
@@ -9,8 +9,9 @@
 import 'dart:math';
 
 import 'package:change_case/change_case.dart';
+import 'package:code_assets/code_assets.dart';
+import 'package:hooks/hooks.dart';
 import 'package:logging/logging.dart';
-import 'package:native_assets_cli/code_assets.dart';
 
 import '../native_toolchain/android_ndk.dart';
 import '../native_toolchain/cmake.dart';