+++ test/builder/cmake_builder_cross_android_test.dart
@@ -4,6 +4,8 @@
 
 import 'dart:io';
 
+import 'package:code_assets/code_assets.dart';
+import 'package:hooks/hooks.dart';
 import 'package:native_toolchain_cmake/native_toolchain_cmake.dart';
 import 'package:native_toolchain_cmake/src/utils/run_process.dart';
 import 'package:test/test.dart';
@@ -108,7 +110,6 @@ Future<Uri> buildLib(
       packageName: name,
       packageRoot: tempUri,
       outputFile: tempUri.resolve('output.json'),
-      outputDirectory: tempUri,
       outputDirectoryShared: tempUriShared,
     )
     ..config.setupBuild(linkingEnabled: false)