+++ test/builder/cmake_builder_cross_ios_test.dart
@@ -10,6 +10,8 @@ library;
 
 import 'dart:io';
 
+import 'package:code_assets/code_assets.dart';
+import 'package:hooks/hooks.dart';
 import 'package:native_toolchain_cmake/native_toolchain_cmake.dart';
 import 'package:native_toolchain_cmake/src/utils/run_process.dart';
 import 'package:test/test.dart';
@@ -53,7 +55,6 @@ void main() {
             packageName: name,
             packageRoot: tempUri,
             outputFile: tempUri.resolve('output.json'),
-            outputDirectory: tempUri,
             outputDirectoryShared: tempUri2,
           )
           ..config.setupBuild(linkingEnabled: false)
@@ -173,7 +174,6 @@ Future<Uri> buildLib(
       packageName: name,
       packageRoot: tempUri,
       outputFile: tempUri.resolve('output.json'),
-      outputDirectory: tempUri,
       outputDirectoryShared: tempUri2,
     )
     ..config.setupBuild(linkingEnabled: false)